+++ internal/app/admin/http/handler_test.go
@@ -3,30 +3,17 @@ package handler
 import (
 	"bytes"
 	"context"
-	"encoding/json"
-	"io/ioutil"
+	v2 "github.com/envoyproxy/go-control-plane/envoy/api/v2"
+	gcp "github.com/envoyproxy/go-control-plane/pkg/cache/v2"
+	"github.com/envoyproxy/xds-relay/internal/app/transport"
+	"github.com/envoyproxy/xds-relay/internal/pkg/log"
+	bootstrapv1 "github.com/envoyproxy/xds-relay/pkg/api/bootstrap/v1"
+	"github.com/golang/protobuf/ptypes"
+	"github.com/golang/protobuf/ptypes/any"
+	"github.com/stretchr/testify/assert"
 	"net/http"
 	"net/http/httptest"
 	"testing"
-	"time"
-
-	"github.com/envoyproxy/xds-relay/internal/app/transport"
-	"github.com/envoyproxy/xds-relay/internal/app/upstream"
-
-	"github.com/envoyproxy/xds-relay/internal/pkg/log"
-	"github.com/envoyproxy/xds-relay/internal/pkg/stats"
-
-	"github.com/envoyproxy/xds-relay/internal/app/mapper"
-	"github.com/envoyproxy/xds-relay/internal/app/orchestrator"
-	"github.com/golang/protobuf/ptypes"
-	"github.com/uber-go/tally"
-
-	v2 "github.com/envoyproxy/go-control-plane/envoy/api/v2"
-	corev2 "github.com/envoyproxy/go-control-plane/envoy/api/v2/core"
-	gcp "github.com/envoyproxy/go-control-plane/pkg/cache/v2"
-	bootstrapv1 "github.com/envoyproxy/xds-relay/pkg/api/bootstrap/v1"
-	"github.com/golang/protobuf/ptypes/any"
-	"github.com/stretchr/testify/assert"
 )
 
 func TestAdminServer_DefaultHandler(t *testing.T) {
@@ -92,213 +79,6 @@ func TestAdminServer_ConfigDumpHandler(t *testing.T) {
 `,
 			rr.Body.String())
 	}
-}
-
-func TestAdminServer_CacheDumpHandler(t *testing.T) {
-	ctx := context.Background()
-	mapper := mapper.NewMock(t)
-	upstreamResponseChannel := make(chan *v2.DiscoveryResponse)
-	mockScope := tally.NewTestScope("mock_orchestrator", make(map[string]string))
-	client := upstream.NewMock(
-		ctx,
-		upstream.CallOptions{SendTimeout: time.Second},
-		nil,
-		upstreamResponseChannel,
-		nil,
-		nil,
-		nil,
-		func(m interface{}) error { return nil },
-		stats.NewMockScope("mock"),
-	)
-	orchestrator := orchestrator.NewMock(t, mapper, client, mockScope)
-	assert.NotNil(t, orchestrator)
-
-	reqNode := corev2.Node{
-		Id:      "test-1",
-		Cluster: "test-prod",
-	}
-	gcpReq := gcp.Request{
-		TypeUrl: "type.googleapis.com/envoy.api.v2.Listener",
-		Node:    &reqNode,
-	}
-	respChannel, cancelWatch := orchestrator.CreateWatch(transport.NewRequestV2(&gcpReq))
-	assert.NotNil(t, respChannel)
-
-	listener := &v2.Listener{
-		Name: "lds resource",
-	}
-	listenerAny, err := ptypes.MarshalAny(listener)
-	assert.NoError(t, err)
-	resp := v2.DiscoveryResponse{
-		VersionInfo: "1",
-		TypeUrl:     "type.googleapis.com/envoy.api.v2.Listener",
-		Resources: []*any.Any{
-			listenerAny,
-		},
-	}
-	upstreamResponseChannel <- &resp
-	gotResponse := <-respChannel.GetChannel().V2
-	gotDiscoveryResponse, err := gotResponse.GetDiscoveryResponse()
-	assert.NoError(t, err)
-	assert.Equal(t, &resp, gotDiscoveryResponse)
-
-	req, err := http.NewRequest("GET", "/cache/test_lds", nil)
-	assert.NoError(t, err)
-
-	rr := httptest.NewRecorder()
-	handler := cacheDumpHandler(&orchestrator)
-
-	handler.ServeHTTP(rr, req)
-	assert.Equal(t, http.StatusOK, rr.Code)
-
-	var actualResponse map[string]interface{}
-	err = json.Unmarshal(rr.Body.Bytes(), &actualResponse)
-	assert.NoError(t, err)
-
-	filecontentsLds, err := ioutil.ReadFile("testdata/entire_cachev2_lds.json")
-	assert.NoError(t, err)
-	var expectedLdsResponse map[string]interface{}
-	err = json.Unmarshal(filecontentsLds, &expectedLdsResponse)
-	assert.NoError(t, err)
-
-	actualCacheResponse := actualResponse["Cache"].([]interface{})
-	assert.Equal(t, len(actualCacheResponse), 1)
-
-	assert.Equal(t, expectedLdsResponse["Key"], actualCacheResponse[0].(map[string]interface{})["Key"])
-	assert.Equal(t, expectedLdsResponse["Resp"], actualCacheResponse[0].(map[string]interface{})["Resp"])
-	cancelWatch()
-}
-
-func TestAdminServer_CacheDumpHandler_NotFound(t *testing.T) {
-	ctx := context.Background()
-	mapper := mapper.NewMock(t)
-	upstreamResponseChannel := make(chan *v2.DiscoveryResponse)
-	mockScope := tally.NewTestScope("mock_orchestrator", make(map[string]string))
-	client := upstream.NewMock(
-		ctx,
-		upstream.CallOptions{SendTimeout: time.Second},
-		nil,
-		nil,
-		nil,
-		nil,
-		upstreamResponseChannel,
-		func(m interface{}) error { return nil },
-		stats.NewMockScope("mock"),
-	)
-	orchestrator := orchestrator.NewMock(t, mapper, client, mockScope)
-	assert.NotNil(t, orchestrator)
-
-	req, err := http.NewRequest("GET", "/cache/cds", nil)
-	assert.NoError(t, err)
-
-	rr := httptest.NewRecorder()
-	handler := cacheDumpHandler(&orchestrator)
-
-	handler.ServeHTTP(rr, req)
-	assert.Equal(t, http.StatusOK, rr.Code)
-	assert.Equal(t, "", rr.Body.String())
-}
-
-func TestAdminServer_ClearCacheHandler(t *testing.T) {
-	ctx := context.Background()
-	mapper := mapper.NewMock(t)
-	upstreamResponseChannel := make(chan *v2.DiscoveryResponse)
-	mockScope := tally.NewTestScope("mock_orchestrator", make(map[string]string))
-	client := upstream.NewMock(
-		ctx,
-		upstream.CallOptions{SendTimeout: time.Second},
-		nil,
-		upstreamResponseChannel,
-		nil,
-		nil,
-		nil,
-		func(m interface{}) error { return nil },
-		stats.NewMockScope("mock"),
-	)
-	orchestrator := orchestrator.NewMock(t, mapper, client, mockScope)
-	assert.NotNil(t, orchestrator)
-
-	reqNode := corev2.Node{
-		Id:      "test-1",
-		Cluster: "test-prod",
-	}
-	gcpReq := gcp.Request{
-		TypeUrl: "type.googleapis.com/envoy.api.v2.Listener",
-		Node:    &reqNode,
-	}
-	ldsRespChannel, cancelWatch := orchestrator.CreateWatch(transport.NewRequestV2(&gcpReq))
-	assert.NotNil(t, ldsRespChannel)
-
-	listener := &v2.Listener{
-		Name: "lds resource",
-	}
-	listenerAny, err := ptypes.MarshalAny(listener)
-	assert.NoError(t, err)
-	resp := v2.DiscoveryResponse{
-		VersionInfo: "1",
-		TypeUrl:     "type.googleapis.com/envoy.api.v2.Listener",
-		Resources: []*any.Any{
-			listenerAny,
-		},
-	}
-	upstreamResponseChannel <- &resp
-	gotResponse := <-ldsRespChannel.GetChannel().V2
-	gotDiscoveryResponse, err := gotResponse.GetDiscoveryResponse()
-	assert.NoError(t, err)
-	assert.Equal(t, &resp, gotDiscoveryResponse)
-
-	// Assert cache has one entry before clearing
-	cacheKeys, err := orchestrator.GetDownstreamAggregatedKeys()
-	assert.NoError(t, err)
-	assert.Equal(t, len(cacheKeys), 1)
-
-	req, err := http.NewRequest("POST", "/clear_cache/test_lds", nil)
-	assert.NoError(t, err)
-
-	rr := httptest.NewRecorder()
-	handler := clearCacheHandler(&orchestrator)
-
-	handler.ServeHTTP(rr, req)
-	assert.Equal(t, http.StatusOK, rr.Code)
-
-	// Assert cache has zero entries after clearing
-	cacheKeys, err = orchestrator.GetDownstreamAggregatedKeys()
-	assert.NoError(t, err)
-	assert.Equal(t, len(cacheKeys), 0)
-
-	cancelWatch()
-}
-
-func TestAdminServer_ClearCacheHandler_NotFound(t *testing.T) {
-	ctx := context.Background()
-	mapper := mapper.NewMock(t)
-	upstreamResponseChannel := make(chan *v2.DiscoveryResponse)
-	mockScope := tally.NewTestScope("mock_orchestrator", make(map[string]string))
-	client := upstream.NewMock(
-		ctx,
-		upstream.CallOptions{SendTimeout: time.Second},
-		nil,
-		nil,
-		nil,
-		nil,
-		upstreamResponseChannel,
-		func(m interface{}) error { return nil },
-		stats.NewMockScope("mock"),
-	)
-	orchestrator := orchestrator.NewMock(t, mapper, client, mockScope)
-	assert.NotNil(t, orchestrator)
-
-	req, err := http.NewRequest("POST", "/clear_cache/cds", nil)
-	assert.NoError(t, err)
-
-	rr := httptest.NewRecorder()
-	handler := clearCacheHandler(&orchestrator)
-
-	handler.ServeHTTP(rr, req)
-	assert.Equal(t, http.StatusOK, rr.Code)
-	assert.Equal(t,
-		"{\n  \"Message\": \"unable to delete entry for nonexistent key: cds\"\n}",
-		rr.Body.String())
 }
 
 func TestGetParam(t *testing.T) {
