+++ internal/app/cache/cache.go
@@ -30,7 +30,7 @@ type Cache interface {
 	DeleteRequest(key string, req transport.Request) error
 
 	// DeleteKey removes the given key from cache.
-	DeleteKey(key string) (Resource, error)
+	DeleteKey(key string) error
 
 	// GetReadOnlyCache returns a copy of the cache that only exposes read-only methods in its interface.
 	GetReadOnlyCache() ReadOnlyCache
@@ -243,23 +243,18 @@ func (c *cache) DeleteRequest(key string, req transport.Request) error {
 	return nil
 }
 
-func (c *cache) DeleteKey(key string) (Resource, error) {
+func (c *cache) DeleteKey(key string) error {
 	c.cacheMu.Lock()
 	defer c.cacheMu.Unlock()
 	metrics.CacheDeleteKeySubscope(c.scope, key).Counter(metrics.CacheDeleteKeyAttempt).Inc(1)
-	value, found := c.cache.Get(key)
+	_, found := c.cache.Get(key)
 	if !found {
 		metrics.CacheDeleteKeySubscope(c.scope, key).Counter(metrics.CacheDeleteKeyError).Inc(1)
-		return Resource{}, fmt.Errorf(fmt.Sprintf("unable to delete entry for nonexistent key: %s", key))
-	}
-	resource, ok := value.(Resource)
-	if !ok {
-		metrics.CacheDeleteRequestSubscope(c.scope, key).Counter(metrics.CacheDeleteKeyError).Inc(1)
-		return Resource{}, fmt.Errorf(fmt.Sprintf("unable to cast cache value to type resource for key: %s", key))
+		return fmt.Errorf(fmt.Sprintf("unable to delete entry for nonexistent key: %s", key))
 	}
 	c.cache.Remove(key)
 	metrics.CacheDeleteKeySubscope(c.scope, key).Counter(metrics.CacheDeleteKeySuccess).Inc(1)
-	return resource, nil
+	return nil
 }
 
 func (r *Resource) isExpired(currentTime time.Time) bool {
